pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "user_embeddings"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bokeh>=3.7.2",
    "dask[distributed]>=2025.3.0",
    "dotenv>=0.9.9",
    "dtale>=3.17.0",
    "faiss-cpu>=1.10.0",
    "google-genai>=1.10.0",
    "httpx>=0.28.1",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.6",
    "json-repair>=0.41.0",
    "kaleido==0.2.1",
    "nbformat>=5.10.4",
    "openai>=1.73.0",
    "polars>=1.27.1",
    "pyarrow>=19.0.1",
    "pytest>=8.3.5",
    "torch>=2.6.0",
    "tqdm>=4.67.1",
    "transformers>=4.51.2",
]

[tool.setuptools.packages.find]
where = ["src"] 

[dependency-groups]
dev = [
    "ruff>=0.11.5",
]

[tool.ruff]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# https://docs.astral.sh/ruff/rules/
# Add explicit rules as needed
select = ["E4", "E7", "E9", "F", "W"]

# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.13
target-version = "py313"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"
